Three tab navigator converted into a general Navigator
